Moves deep link listener logic to its own file.
